pyproject.toml
[project]
name = "ai-animation-redraw"
version = "0.1.0"
description = "AI手书复刻 - Redraw animation clips with AI-generated characters"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "pillow>=10.4.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=24.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
